+++ src/model/key_positions.rs
@@ -70,7 +70,7 @@ pub struct Columns(Vec<Column>);
 
 impl Columns {
     fn from_config(config: &FingerCluster) -> Self {
-        const CURVATURE_HEIGHT: f64 = Switch::TOP_HEIGHT;
+        const CURVATURE_HEIGHT: f64 = Switch::TOP_HEIGHT as f64;
 
         let key_distance: PositiveDVec2 = (&config.key_distance).into();
 
@@ -190,7 +190,7 @@ pub struct ThumbKeys(Vec<DAffine3>);
 
 impl ThumbKeys {
     fn from_config(config: &ThumbCluster) -> Self {
-        const CURVATURE_HEIGHT: f64 = Switch::TOP_HEIGHT;
+        const CURVATURE_HEIGHT: f64 = Switch::TOP_HEIGHT as f64;
 
         let curvature_angle = config.curvature_angle.to_radians();
         let cluster_rotation = DQuat::from_euler(
